% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_calc.R
\name{cv_x}
\alias{cv_x}
\title{Calculate coefficient of variation (standard deviation/mean)}
\usage{
cv_x(x)
}
\arguments{
\item{x}{numeric vector}
}
\value{
coefficient of variation (CV) value
}
\description{
Calculate coefficient of variation (standard deviation/mean)
}
\examples{
#' #' \dontrun{
cv_x(c(0,1,1,3,0,5))
}
}
