% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_calc.R
\name{disp_x}
\alias{disp_x}
\title{Calculate dispersion (variance/mean)}
\usage{
disp_x(x)
}
\arguments{
\item{x}{numeric vector}
}
\value{
dispersion value
}
\description{
Calculate dispersion (variance/mean)
}
\examples{
#' \dontrun{
disp_x(c(0,1,1,3,0,5))
}
}
