% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_calc.R
\name{humpMean}
\alias{humpMean}
\title{humpMean: Mean of a numeric vector x restricted to only values x > min}
\usage{
humpMean(x, min = 0)
}
\arguments{
\item{x}{numeric vector}

\item{min}{numeric value}
}
\value{
mean of entries of x greter than min (numeric)
}
\description{
humpMean: Mean of a numeric vector x restricted to only values x > min
}
\examples{
\dontrun{
humpMean(c(0,1,1,3,0,5), min=0)
}
}
